+++ pkg/ring/kv/dynamodb/client.go
@@ -15,12 +15,17 @@ import (
 	"github.com/cortexproject/cortex/pkg/util/backoff"
 )
 
+const (
+	maxCasRetries = 10 // max retries in CAS operation
+)
+
 // Config to create a ConsulClient
 type Config struct {
 	Region         string        `yaml:"region"`
 	TableName      string        `yaml:"table_name"`
 	TTL            time.Duration `yaml:"ttl"`
 	PullerSyncTime time.Duration `yaml:"puller_sync_time"`
+	MaxCasRetries  int           `yaml:"max_cas_retries"`
 }
 
 type Client struct {
@@ -47,6 +52,7 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet, prefix string) {
 	f.StringVar(&cfg.TableName, prefix+"dynamodb.table-name", "", "Table name to use on dynamodb.")
 	f.DurationVar(&cfg.TTL, prefix+"dynamodb.ttl-time", 0, "Time to expire items on dynamodb.")
 	f.DurationVar(&cfg.PullerSyncTime, prefix+"dynamodb.puller-sync-time", 60*time.Second, "Time to refresh local ring with information on dynamodb.")
+	f.IntVar(&cfg.MaxCasRetries, prefix+"dynamodb.max-cas-retries", maxCasRetries, "Maximum number of retries for DDB KV CAS.")
 }
 
 func NewClient(cfg Config, cc codec.Codec, logger log.Logger, registerer prometheus.Registerer) (*Client, error) {
@@ -60,7 +66,7 @@ func NewClient(cfg Config, cc codec.Codec, logger log.Logger, registerer prometh
 	backoffConfig := backoff.Config{
 		MinBackoff: 1 * time.Second,
 		MaxBackoff: cfg.PullerSyncTime,
-		MaxRetries: 0,
+		MaxRetries: cfg.MaxCasRetries,
 	}
 
 	c := &Client{
@@ -125,6 +131,7 @@ func (c *Client) Delete(ctx context.Context, key string) error {
 func (c *Client) CAS(ctx context.Context, key string, f func(in interface{}) (out interface{}, retry bool, err error)) error {
 	bo := backoff.New(ctx, c.backoffConfig)
 	for bo.Ongoing() {
+		c.ddbMetrics.dynamodbCasAttempts.Inc()
 		resp, _, err := c.kv.Query(ctx, dynamodbKey{primaryKey: key}, false)
 		if err != nil {
 			level.Error(c.logger).Log("msg", "error cas query", "key", key, "err", err)
@@ -183,6 +190,12 @@ func (c *Client) CAS(ctx context.Context, key string, f func(in interface{}) (ou
 
 		if len(putRequests) > 0 || len(deleteRequests) > 0 {
 			return c.kv.Batch(ctx, putRequests, deleteRequests)
+		}
+
+		if len(putRequests) == 0 && len(deleteRequests) == 0 {
+			// no change detected, retry
+			bo.Wait()
+			continue
 		}
 
 		return nil