+++ pkg/storage/tsdb/bucketindex/loader.go
@@ -96,14 +96,21 @@ func (l *Loader) GetIndex(ctx context.Context, userID string) (*Index, Status, e
 	if entry := l.indexes[userID]; entry != nil {
 		idx := entry.index
 		err := entry.err
+		ss := entry.syncStatus
 		l.indexesMx.RUnlock()
 
 		// We don't check if the index is stale because it's the responsibility
 		// of the background job to keep it updated.
 		entry.requestedAt.Store(time.Now().Unix())
-		return idx, entry.syncStatus, err
+		return idx, ss, err
 	}
 	l.indexesMx.RUnlock()
+
+	ss, err := ReadSyncStatus(ctx, l.bkt, userID, l.logger)
+
+	if err != nil {
+		level.Warn(l.logger).Log("msg", "unable to read bucket index status", "user", userID, "err", err)
+	}
 
 	startTime := time.Now()
 	l.loadAttempts.Inc()
@@ -111,7 +118,7 @@ func (l *Loader) GetIndex(ctx context.Context, userID string) (*Index, Status, e
 	if err != nil {
 		// Cache the error, to avoid hammering the object store in case of persistent issues
 		// (eg. corrupted bucket index or not existing).
-		l.cacheIndex(userID, nil, UnknownStatus, err)
+		l.cacheIndex(userID, nil, ss, err)
 
 		if errors.Is(err, ErrIndexNotFound) {
 			level.Warn(l.logger).Log("msg", "bucket index not found", "user", userID)
@@ -124,13 +131,7 @@ func (l *Loader) GetIndex(ctx context.Context, userID string) (*Index, Status, e
 			level.Error(l.logger).Log("msg", "unable to load bucket index", "user", userID, "err", err)
 		}
 
-		return nil, UnknownStatus, err
-	}
-
-	ss, err := ReadSyncStatus(ctx, l.bkt, userID, l.logger)
-
-	if err != nil {
-		level.Warn(l.logger).Log("msg", "unable to read bucket index status", "user", userID, "err", err)
+		return nil, ss, err
 	}
 
 	// Cache the index.
@@ -203,6 +204,16 @@ func (l *Loader) updateCachedIndex(ctx context.Context, userID string) {
 
 	l.loadAttempts.Inc()
 	startTime := time.Now()
+	ss, err := ReadSyncStatus(ctx, l.bkt, userID, l.logger)
+	if err != nil {
+		level.Warn(l.logger).Log("msg", "unable to read bucket index status", "user", userID, "err", err)
+	}
+
+	// Update Sync Status
+	l.indexesMx.Lock()
+	l.indexes[userID].syncStatus = ss
+	l.indexesMx.Unlock()
+
 	idx, err := ReadIndex(readCtx, l.bkt, userID, l.cfgProvider, l.logger)
 	if err != nil && !errors.Is(err, ErrIndexNotFound) && !errors.Is(err, bucket.ErrCustomerManagedKeyAccessDenied) {
 		l.loadFailures.Inc()
@@ -212,7 +223,7 @@ func (l *Loader) updateCachedIndex(ctx context.Context, userID string) {
 
 	l.loadDuration.Observe(time.Since(startTime).Seconds())
 
-	// We cache it either it was successfully refreshed or wasn't found. An use case for caching the ErrIndexNotFound
+	// We cache it either it was successfully refreshed,  wasn't found or when is a CMK error. An use case for caching the ErrIndexNotFound
 	// is when a tenant has rules configured but hasn't started remote writing yet. Rules will be evaluated and
 	// bucket index loaded by the ruler.
 	l.indexesMx.Lock()