+++ pkg/storegateway/chunk_bytes_pool.go
@@ -10,8 +10,7 @@ type chunkBytesPool struct {
 	pool *pool.BucketedBytes
 
 	// Metrics.
-	requestedBytes prometheus.Counter
-	returnedBytes  prometheus.Counter
+	poolByteStats *prometheus.CounterVec
 }
 
 func newChunkBytesPool(minBucketSize, maxBucketSize int, maxChunkPoolBytes uint64, reg prometheus.Registerer) (*chunkBytesPool, error) {
@@ -22,14 +21,10 @@ func newChunkBytesPool(minBucketSize, maxBucketSize int, maxChunkPoolBytes uint6
 
 	return &chunkBytesPool{
 		pool: upstream,
-		requestedBytes: promauto.With(reg).NewCounter(prometheus.CounterOpts{
-			Name: "cortex_bucket_store_chunk_pool_requested_bytes_total",
-			Help: "Total bytes requested to chunk bytes pool.",
-		}),
-		returnedBytes: promauto.With(reg).NewCounter(prometheus.CounterOpts{
-			Name: "cortex_bucket_store_chunk_pool_returned_bytes_total",
-			Help: "Total bytes returned by the chunk bytes pool.",
-		}),
+		poolByteStats: promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+			Name: "cortex_bucket_store_chunk_pool_operation_bytes_total",
+			Help: "Total bytes number of bytes pooled by operation.",
+		}, []string{"operation", "stats"}),
 	}, nil
 }
 
@@ -39,12 +34,14 @@ func (p *chunkBytesPool) Get(sz int) (*[]byte, error) {
 		return buffer, err
 	}
 
-	p.requestedBytes.Add(float64(sz))
-	p.returnedBytes.Add(float64(cap(*buffer)))
+	p.poolByteStats.WithLabelValues("get", "requested").Add(float64(sz))
+	p.poolByteStats.WithLabelValues("get", "cap").Add(float64(cap(*buffer)))
 
 	return buffer, err
 }
 
 func (p *chunkBytesPool) Put(b *[]byte) {
+	p.poolByteStats.WithLabelValues("put", "len").Add(float64(len(*b)))
+	p.poolByteStats.WithLabelValues("put", "cap").Add(float64(cap(*b)))
 	p.pool.Put(b)
 }