+++ pkg/storegateway/chunk_bytes_pool_test.go
@@ -19,20 +19,23 @@ func TestChunkBytesPool_Get(t *testing.T) {
 	reg := prometheus.NewPedanticRegistry()
 	p, err := newChunkBytesPool(cortex_tsdb.ChunkPoolDefaultMinBucketSize, cortex_tsdb.ChunkPoolDefaultMaxBucketSize, 0, reg)
 	require.NoError(t, err)
-
+	testBytes := []byte("test")
 	_, err = p.Get(store.EstimatedMaxChunkSize - 1)
 	require.NoError(t, err)
 
-	_, err = p.Get(store.EstimatedMaxChunkSize + 1)
+	b, err := p.Get(store.EstimatedMaxChunkSize + 1)
 	require.NoError(t, err)
 
-	assert.NoError(t, testutil.GatherAndCompare(reg, bytes.NewBufferString(fmt.Sprintf(`
-		# HELP cortex_bucket_store_chunk_pool_requested_bytes_total Total bytes requested to chunk bytes pool.
-		# TYPE cortex_bucket_store_chunk_pool_requested_bytes_total counter
-		cortex_bucket_store_chunk_pool_requested_bytes_total %d
+	*b = append(*b, testBytes...)
 
-		# HELP cortex_bucket_store_chunk_pool_returned_bytes_total Total bytes returned by the chunk bytes pool.
-		# TYPE cortex_bucket_store_chunk_pool_returned_bytes_total counter
-		cortex_bucket_store_chunk_pool_returned_bytes_total %d
-	`, store.EstimatedMaxChunkSize*2, store.EstimatedMaxChunkSize*3))))
+	p.Put(b)
+
+	assert.NoError(t, testutil.GatherAndCompare(reg, bytes.NewBufferString(fmt.Sprintf(`
+		# HELP cortex_bucket_store_chunk_pool_operation_bytes_total Total bytes number of bytes pooled by operation.
+		# TYPE cortex_bucket_store_chunk_pool_operation_bytes_total counter
+		cortex_bucket_store_chunk_pool_operation_bytes_total{operation="get",stats="cap"} %d
+		cortex_bucket_store_chunk_pool_operation_bytes_total{operation="get",stats="requested"} %d
+		cortex_bucket_store_chunk_pool_operation_bytes_total{operation="put",stats="cap"} %d
+		cortex_bucket_store_chunk_pool_operation_bytes_total{operation="put",stats="len"} %d
+	`, store.EstimatedMaxChunkSize*3, store.EstimatedMaxChunkSize*2, store.EstimatedMaxChunkSize*2, len(testBytes)))))
 }