+++ pkg/util/validation/notifications_limit_flag.go
@@ -11,6 +11,7 @@ import (
 
 var allowedIntegrationNames = []string{
 	"webhook", "email", "pagerduty", "opsgenie", "wechat", "slack", "victorops", "pushover", "sns", "telegram", "discord", "webex",
+	"msteams",
 }
 
 type NotificationRateLimitMap map[string]float64